Graph of temperature across a range
Hey  added the "show graph across a temperature range" feature. the graph is a histogram.
we can make further changes as in making the histogram match the color of the temperatures, etc but this seems satisfactory but needs to be tested

Check the logic and I have set the range such that the latest 5 temperature readings are used to draw the histogram